Removed the heartbeatproducer.  Issues #53 and #27.
- Removed all references to fedmsg-status
- Removed all references to have_pulses
- Removed all references to the HeartbeatProducer

The heartbeat producer was ill-conceived.  It was designed to be a check for
nagios, but services like "bodhi" emit messages without a hub, so the heartbeat
producer is never run.  The same is true for mediawiki, fas, tagger, scm, .. and
everything.  The only service that ever had one was fedmsg-relay.

Removing for now until we can more easily see a better solution.