Protect against NotImplementedError.
In fedmsg-tail, we attempt to extract the relevant usernames and
packages from a message in order to compare them against a list of
usernames or packages that the caller has requested.  However, it is
almost never the case that the caller is requesting such a thing.

This will 1) save time on average and 2) protect the caller in the case
where there is a NotImplementedError coming from one of the message
processors.