A second take on the 'conglomerate' API.
The first approach in #278 didn't provide enough support for what I
ended up needing to implement in fedmsg_meta_fedora_infrastructure.

The gist here is that fedmsg core will provide a ``BaseConglomerator``
class that plugins will extend.  It is an abstract base class, and
children will have to override three different methods for it to work
properly.

I already have a functioning Conglomerator for some bodhi messages that
I'll submit as a PR soon.