Improve git hook to only report on new commits
Consider this git graph:

```
* commit A <-- HEAD
|
* merge commit B
| \
| * commit C
| |
| * commit D
| |
| * merge commit E
| | \
| | * commit F
| | |
| | * commit G
| | |
| | |
* | | commit H  <-- BASE
| | |
| * | commit I
|/ /
* | commit J
| |
| * commit K
|/
* commit L
|
* commit M
|
* commit N
```

If the person pushing was updating master from ``BASE`` to ``HEAD``, the
original implementation sought out all parents of the ``HEAD`` commit
recursively, and the only stop-condition was when ``BASE`` was found for that
branch of the recursion tree.  In the example above, the two branches that
branch out from and back into the right side *do not have* ``BASE`` in their
history. For them, the original algorithm would spider all the way back to the
first commit.

The current implementation is not right either.  It only stops once it finds a
commit that has already been seen (preventing cycles) but it won't prevent this
scenario from publishing  messages about commits J, L, M, and N.

The implementation in this commit works like this:

- First, calculate all the ``ancestors`` of the ``BASE`` rev, including the
  ``BASE`` rev.  In the example above, that would be H, J, L, M, and N.
- Then, traverse recursively backwards from the ``HEAD`` rev.  Stop recursion
  for that branch of the call tree when a parent is found that is in the
  ``ancestors`` list that we previously computed.  This would end up yielding
  commits A, B, C, D, E, F, G, I, K.  (The compliment of the ``ancestors``
  list).