No more infinite loops.
We hit an infinite loop in the git hook again tonight which made for a *lot* of
traffic on the bus.

![](http://i.imgur.com/XRSzIh7.gif)

This rolls back the functionality to what we used to have before the move to
gitolite3/rhel7.  It translates as "if someone pushes 5 commits, only publish
about the latest one".