Move the daemon dependency to the consumers extra
As it turns out, the dependencies of the packaged version of fedmsg were
incorrect, in that the commands grouped into 'consumers' are the ones which
require the daemon package, while no commands in 'commands' do. This is still
true even if one considers that the python-fedmsg-commands RPM package includes
fedmsg-collectd.