Change test framework to allow testing None values.
Previously, if the ``expected_{value}`` on a test was set to ``None``, then
that test would be skipped.  However, that didn't allow us to test that a value
returned *really was None* or not.

This change introduces the use of a sentinel value called ``Unspecified`` that
we use instead of ``None`` so that we can now set a test to ``None`` over in
fedmsg_meta_fedora_infrastructure and find out if the actual value returned by
our code really was None or not.