Add a hmac backend to sign and validate message send over the fedmsg bus
This backend should make it easier for people to use fedmsg for smaller
deployment than the one used by Debian or the Fedora infrastructure and
where it is easier to keep the secret used... secret.
It should also make it easier to run fedmsg for development while still
having signed messages.