Use cryptography and pyOpenSSL instead of M2Crypto
M2Crypto is Python 2 only and is in maintenance mode. Although pyOpenSSL
is _also_ in maintenance mode, it supports Python 3 and relies on
cryptography for its OpenSSL bindings. Currently the only thing
cryptography is missing (that pyOpenSSL has) is certificate chain
validation. Once that's available in cryptography, we can drop
pyOpenSSL.

Unfortunately, we still need to support M2Crypto because the current
versions of cryptography and pyOpenSSL in EL7 is below 1.6 and 16.1.0
respectively.

This patch uses cryptography and pyOpenSSL if available and falls back
to M2Crypto.

fixes #411

Signed-off-by: Jeremy Cline <jeremy@jcline.org>