Enable CRLs to be unspecified
Prior to this commit, it was required that a CRL was specified. This was
fine with the old M2Crypto code that didn't properly check the CRL for
expiration, but the new crytography-based code will fail if the CRL
is expired. This allows the definition of a CRL in the configuration to
be optional so users don't have to keep a fresh CRL that they may not
actually be using available.

Signed-off-by: Jeremy Cline <jeremy@jcline.org>