Re-organize the fedmsg documentation
The fedmsg documentation has gather a lot of cruft over the years -
there are many files that were not included in the toctree or were no
longer relevant. Additionally, the documentation was split up in
confusing ways (for example, the STOMP configuration was not in the
configuration section).

This commit removes documentation that no longer applies and moves as
much documentation as possible into the Python doc blocks. It defines a
public API. It tries to organize the documentation so that it is easy
to find the section you're interested in.

Signed-off-by: Jeremy Cline <jeremy@jcline.org>