Assert the signature/certificate fields are unicode
This checks the certificate and signature fields in messages to ensure
they're text. If they're not, an error is logged and we do our best to
decode the message as UTF-8. This decoding is a short-term measure until
the users of the APIs have been hunted down and fixed.

This is a cherry-pick from
https://github.com/fedora-infra/fedmsg/pull/456

Signed-off-by: Jeremy Cline <jeremy@jcline.org>