Deserialize message bodies when they're unicode
The validate API should deserialize message bodies of ZMQMessages if
they're six.text_type. Additionally, we'll (for backwards compatibility)
attempt to decode bytes to UTF-8 and log an error since this is not a
safe API and is deprecated.

This is a cherry-pick of https://github.com/fedora-infra/fedmsg/pull/464

Signed-off-by: Jeremy Cline <jeremy@jcline.org>