Make fedmsg replays from datagrepper work in most cases
Message playback was not working because the messages did not pass
validation. The reason they didn't pass validation is datanommer mutates
the original fedmsg. This breaks the x509 signature. This change tries
to undo all the changes datanommer makes, but there are several cases
where it is impossible to know what the correct value for a key is so
some messages might still fail signature validation.

The longer term fix is to change how datanommer works so that it does
not alter the message in any way. Additionally, the fedmsg signature
approach could be made much less fragile.

For those that come after me, I am sorry about all the things I did in
this patch.

fixes #403

Signed-off-by: Jeremy Cline <jeremy@jcline.org>