Strengthen "legacy protection". (#493)
Long ago, we introduced a decorator around the msg2 functions that, in the
event of a `KeyError` would try to gracefully return a default value.  This
protection only kicked in *if* the msg2 function was called with
`legacy=True`... i.e., if it was expecting a legacy msg to be handled.

In practice, we never know if a legacy msg is being handled until its too late.
This change renames the handler and broadens its role to catch exceptions from
any message.  They are logged, and the default value is returned.