Return any return value from ._consume(..).
In a non-zeromq case, it matters what the return value of _consume is.  (It is
a boolean, indicating if the message was successfully handled or not).  This,
in turn, gets used to figure out whether or not to send an ACK back to the
broker.

Here, the return values were being swallowed and turned into `None` values.