Move configuration documentation into the Python module
This moves the documentation for each configuration value into the
module docblock and alters the Sphinx auto-generated documentation to
include the module-level documentation in the user-facing configuration
page. The configuration API docs used to be the place where
configuration load order was documented and users weren't finding it.

The documentation itself is not changed.

Fixes #505

Signed-off-by: Jeremy Cline <jeremy@jcline.org>