handle messages with non-dict bodies
Lots of places in the fedmsg code assume that message['body'] is a dict. However,
it is possible for a producer to send a message with a body that is valid JSON
but does not encode a dict, which would cause validate() to fail. This change
ensures that message['body'] is always a dict.