always enable the MonitoringProducer
moksha records the time it takes a Consumer to process every message
in the Consumer._times list. The MonitoringProducer sends those times
to the monitor socket, and resets the list. If the MonitoringProducer
isn't running, that list grows forever, causing a memory leak. The
default configuration should always enable the MonitoringProducer.